fixing more clang warnings

+++ scilab/modules/boolean/sci_gateway/c/sci_and.c
@@ -43,7 +43,7 @@ int sci_and(char *fname, unsigned long l)
 
     if (!isBooleanType(pvApiCtx, piAddressVarOne))
     {
-        callOverloadFunction(pvApiCtx, 1, fname, l);
+        callOverloadFunction(pvApiCtx, 1, fname, (unsigned int)l);
         return 0;
     }
