+++ scilab/modules/graphics/sci_gateway/c/sci_param3d1.c
@@ -32,8 +32,8 @@ int sci_param3d1( char *fname, unsigned long fname_len )
     static double  ebox_def [6] = { 0, 1, 0, 1, 0, 1};
     double *ebox = ebox_def ;
     static int iflag_def[3] = {1, 2, 4};
-    int iflag[3] , *ifl = NULL;
-    double  alpha_def = 35.0 , theta_def = 45.0 ;
+    int iflag[3], *ifl = NULL;
+    double  alpha_def = 35.0, theta_def = 45.0 ;
     double *alpha = &alpha_def, *theta = &theta_def;
     int m1 = 0, n1 = 0, l1 = 0, m2 = 0, n2 = 0, l2 = 0, m3 = 0, n3 = 0, l3 = 0;
     int m3n = 0, n3n = 0, l3n = 0, m3l = 0, n3l = 0, l3l = 0;
@@ -49,7 +49,7 @@ int sci_param3d1( char *fname, unsigned long fname_len )
 
     if (Rhs <= 0)
     {
-        sci_demo(fname, fname_len);
+        sci_demo(fname, (int)fname_len);
         return 0;
     }
