+++ scilab/modules/graphics/sci_gateway/c/sci_plot2d.c
@@ -66,7 +66,7 @@ int sci_plot2d( char * fname, unsigned long fname_len )
 
     if (Rhs == 0)
     {
-        sci_demo(fname, fname_len);
+        sci_demo(fname, (int)fname_len);
         return 0;
     }
